Cloned skeleton project from: "https://github.com/autobooks/interview-dotnet3".
Added new XUnit project for testing and added project reference to API project.

Added new class "CustomerTests" to write the first red (failing) test. Also added Customer and CustomerController classes to support the first test (these will be moved to the API project in their propper folders when completing the TDD cycle in the Refactor stage.

Updated Startup.cs class Configure method to use IWebHostEnvironment because the original was deprecated.

Completed first red (failing) test for the first requirement which is to retrieve a listing of all customers.

Next I will write the code to make this first test green do that we successfully pass the test.